Put canonicalizeQuery in test utils
