+++ lib/util.ts
@@ -445,60 +445,6 @@ export default class Util
     }
 }
 
-export class Canonicalizer {
-    public blankId: number;
-    public genBlankId() {
-        return "g_" + this.blankId++;
-    }
-
-    public canonicalizeQuery(res: Algebra.Operation) : Algebra.Operation {
-        this.blankId = 0;
-        let nameMapping: { [bLabel: string]: string } = {};
-        return LibUtil.default.mapOperation(res, {
-            'path': (op: Algebra.Path, factory: Factory) => {
-                return {
-                    result: factory.createPath(
-                        this.getNewBlank(op.subject, nameMapping),
-                        op.predicate,
-                        this.getNewBlank(op.object, nameMapping),
-                        this.getNewBlank(op.graph, nameMapping),
-                    ),
-                    recurse: true,
-                };
-            },
-            'pattern': (op: Algebra.Pattern, factory: Factory) => {
-                return {
-                    result: factory.createPattern(
-                        this.getNewBlank(op.subject, nameMapping),
-                        this.getNewBlank(op.predicate, nameMapping),
-                        this.getNewBlank(op.object, nameMapping),
-                        this.getNewBlank(op.graph, nameMapping),
-                    ),
-                    recurse: true,
-                };
-            },
-            'construct': (op: Algebra.Construct, factory) => {
-                // Blank nodes in CONSTRUCT templates must be maintained
-                return {
-                    result: factory.createConstruct(op.input, op.template),
-                    recurse: true,
-                };
-            },
-        });
-    }
-
-    public getNewBlank(term: RDF.Term, nameMapping: {[bLabel: string]: string}): RDF.Term {
-        if (term.termType !== "BlankNode") return term;
-        else {
-            if (nameMapping[term.value]) {
-                return blankNode(nameMapping[term.value]);
-            } else {
-                return blankNode(this.genBlankId());
-            }
-        }
-    }
-}
-
 /**
  * @interface RecurseResult
  * @property {Operation} result - The resulting A.Operation.
