+++ test/sparql-test.ts
@@ -4,8 +4,10 @@ import * as fs from 'fs';
 import * as path from 'path';
 import {translate, toSparql, toSparqlJs} from '../index';
 import Util from './util';
+import canonicalize = Mocha.utils.canonicalize;
 
 const rootJson = 'test/algebra';
+const canon = Util.getCanonicalizerInstance();
 
 // https://www.w3.org/2001/sw/DataAccess/tests/r2#syntax-basic-01
 // https://www.w3.org/2009/sparql/implementations/
@@ -33,7 +35,7 @@ function testPath(fileName: string, testName: string): void
             let expected = JSON.parse(fs.readFileSync(jsonName, 'utf8'));
             let query = toSparql(expected);
             let algebra = Util.objectify(translate(query, { quads: name.endsWith('(quads)') }));
-            expect(algebra).to.deep.equal(expected);
+            expect(canon.canonicalizeQuery(algebra)).to.deep.equal(canon.canonicalizeQuery(expected));
         });
     }
 }
